+++ RELEASE-NOTES.md
@@ -1,8 +1,9 @@
 # Release notes
 ## 3.2
-Released January 2018. See [log](https://github.com/Pathoschild/FluentHttpClient/compare/3.1...3.2).
+The upcoming release. See [log](https://github.com/Pathoschild/FluentHttpClient/compare/3.1...3.2).
 
-* Improved the `Request.WithArguments` method to accept multiple arguments with the same name
+* Added `Request.WithArguments` overload which takes `KeyValuePair` arguments to simplify multiple argument values.
+* Fixed error when passing an object with an indexer property to `WithArguments`.
 
 ## 3.1
 Released 19 September 2017. See [log](https://github.com/Pathoschild/FluentHttpClient/compare/3.0...3.1).